wallet-ext: useExplorerLink hook
* move the creation of the explorer link out of the ExplorerLink component
